LuckyTicets

+++ LuckyTicets/LuckyTicets/Program.cs
@@ -5,19 +5,19 @@ namespace LuckyTicket
     {
         public static void Main()
         {
-            long n = 4;
-            long half = n / 2;
+            int n = 4;
+            int half = n / 2;
             var luckyTickets = new LuckyTicket();
             Console.WriteLine(luckyTickets.Run(half));
         }
         long count = 0;
-        public long Run(long half)
+        public long Run(int half)
         {
             RecursionSeach(half, 0, 0);
             return count;
         }
 
-        public void RecursionSeach(long half, int sumA, int sumB)
+        public void RecursionSeach(int half, int sumA, int sumB)
         {
             if (half == 0)
             {
